refactor: extract summary
